Docker now detachable
Originally, when inside the docker containter the user was unable to escape out of the container with the default key combination. The docker would also kill itself upon detachemnt. Now the user can have the witness running it tmux, pust tmux to the background with  and then enter  to escape the docker container and keep in running in the background.